requirements.txt
fastapi==0.103.1
langchain==0.0.330
python-dotenv==1.0.0
beautifulsoup4==4.12.2
Brotli==1.1.0
exceptiongroup==1.1.3
google-api-core==2.12.0
lxml==4.9.3
numexpr==2.8.7
openapi-schema-pydantic==1.2.4
pdfminer.six==20221105
pip==23.3
pyOpenSSL==23.3.0
PySocks==1.7.1
pyu2f==0.1.5
setuptools==68.0.0
stringcase==1.2.0
wheel==0.41.2
uvicorn==0.24.0
anthropic==0.5.0
pydantic==2.4.2
pydantic-settings==2.0.3
<<<<<<< Updated upstream
nltk==3.8.1
=======
anthropic==0.5.0
bs4==0.0.1
email-validator==2.1.0.post1
httptools==0.6.1
itsdangerous==2.1.2
lxml==4.9.3
orjson==3.9.10
protobuf==4.25.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
python-multipart==0.0.6
sentence-transformers==2.2.2
torch-utils==0.1.2
ujson==5.8.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
>>>>>>> Stashed changes
